+++ Cargo.toml
@@ -114,7 +114,6 @@ regex = "1.7.0"
 reqwest = { version = "0.11.13", default-features = false, features=["cookies", "json", "gzip", "native-tls"] }
 rpassword = "^7.2.0"
 rusqlite = "^0.28.0"
-saffron = "^0.1.0"
 serde = "^1.0.151"
 serde_cbor = { version = "0.12.0-dev", package = "serde_cbor_2" }
 serde_json = "^1.0.91"
