+++ kanidmd/lib/src/filter.rs
@@ -1561,18 +1561,11 @@ mod tests {
 
     #[test]
     fn test_lessthan_entry_filter() {
-        let e: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "userid": ["william"],
-                "uuid": ["db237e8a-0079-4b8c-8a56-593b22aa44d1"],
-                "gidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e = unsafe { entry_init!(
+            ("userid", Value::new_iutf8("william")),
+            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
         let f_t1a = unsafe { filter_resolved!(f_lt("gidnumber", PartialValue::new_uint32(500))) };
         assert!(e.entry_match_no_index(&f_t1a) == false);
@@ -1586,31 +1579,24 @@ mod tests {
 
     #[test]
     fn test_or_entry_filter() {
-        let e: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "userid": ["william"],
-                "uuid": ["db237e8a-0079-4b8c-8a56-593b22aa44d1"],
-                "uidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e = unsafe { entry_init!(
+            ("userid", Value::new_iutf8("william")),
+            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
         let f_t1a = unsafe {
             filter_resolved!(f_or!([
                 f_eq("userid", PartialValue::new_iutf8("william")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1000")),
-            ]))
+                f_eq("gidnumber", PartialValue::Uint32(1000)),
+           ]))
         };
         assert!(e.entry_match_no_index(&f_t1a));
 
         let f_t2a = unsafe {
             filter_resolved!(f_or!([
                 f_eq("userid", PartialValue::new_iutf8("william")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1001")),
+                f_eq("gidnumber", PartialValue::Uint32(1000)),
             ]))
         };
         assert!(e.entry_match_no_index(&f_t2a));
@@ -1618,7 +1604,7 @@ mod tests {
         let f_t3a = unsafe {
             filter_resolved!(f_or!([
                 f_eq("userid", PartialValue::new_iutf8("alice")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1000")),
+                f_eq("gidnumber", PartialValue::Uint32(1000)),
             ]))
         };
         assert!(e.entry_match_no_index(&f_t3a));
@@ -1626,7 +1612,7 @@ mod tests {
         let f_t4a = unsafe {
             filter_resolved!(f_or!([
                 f_eq("userid", PartialValue::new_iutf8("alice")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1001")),
+                f_eq("gidnumber", PartialValue::Uint32(1001)),
             ]))
         };
         assert!(!e.entry_match_no_index(&f_t4a));
@@ -1634,23 +1620,16 @@ mod tests {
 
     #[test]
     fn test_and_entry_filter() {
-        let e: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "userid": ["william"],
-                "uuid": ["db237e8a-0079-4b8c-8a56-593b22aa44d1"],
-                "uidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e = unsafe { entry_init!(
+            ("userid", Value::new_iutf8("william")),
+            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
         let f_t1a = unsafe {
             filter_resolved!(f_and!([
                 f_eq("userid", PartialValue::new_iutf8("william")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1000")),
+                f_eq("gidnumber", PartialValue::Uint32(1000)),
             ]))
         };
         assert!(e.entry_match_no_index(&f_t1a));
@@ -1658,7 +1637,7 @@ mod tests {
         let f_t2a = unsafe {
             filter_resolved!(f_and!([
                 f_eq("userid", PartialValue::new_iutf8("william")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1001")),
+                f_eq("gidnumber", PartialValue::Uint32(1001)),
             ]))
         };
         assert!(!e.entry_match_no_index(&f_t2a));
@@ -1666,7 +1645,7 @@ mod tests {
         let f_t3a = unsafe {
             filter_resolved!(f_and!([
                 f_eq("userid", PartialValue::new_iutf8("alice")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1000")),
+                f_eq("gidnumber", PartialValue::Uint32(1000)),
             ]))
         };
         assert!(!e.entry_match_no_index(&f_t3a));
@@ -1674,7 +1653,7 @@ mod tests {
         let f_t4a = unsafe {
             filter_resolved!(f_and!([
                 f_eq("userid", PartialValue::new_iutf8("alice")),
-                f_eq("uidnumber", PartialValue::new_iutf8("1001")),
+                f_eq("gidnumber", PartialValue::Uint32(1001)),
             ]))
         };
         assert!(!e.entry_match_no_index(&f_t4a));
@@ -1682,18 +1661,11 @@ mod tests {
 
     #[test]
     fn test_not_entry_filter() {
-        let e1: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "userid": ["william"],
-                "uuid": ["db237e8a-0079-4b8c-8a56-593b22aa44d1"],
-                "uidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e1 = unsafe { entry_init!(
+            ("userid", Value::new_iutf8("william")),
+            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
         let f_t1a =
             unsafe { filter_resolved!(f_andnot(f_eq("userid", PartialValue::new_iutf8("alice")))) };
@@ -1707,64 +1679,36 @@ mod tests {
 
     #[test]
     fn test_nested_entry_filter() {
-        let e1: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "class": ["person"],
-                "uuid": ["db237e8a-0079-4b8c-8a56-593b22aa44d1"],
-                "uidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e1 = unsafe { entry_init!(
+            ("class", CLASS_PERSON.clone()),
+            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
-        let e2: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "class": ["person"],
-                "uuid": ["4b6228ab-1dbe-42a4-a9f5-f6368222438e"],
-                "uidnumber": ["1001"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e2 = unsafe { entry_init!(
+            ("class", CLASS_PERSON.clone()),
+            ("uuid", Value::Uuid(uuid::uuid!("4b6228ab-1dbe-42a4-a9f5-f6368222438e"))),
+            ("gidnumber", Value::Uint32(1001))
+        ).into_sealed_new() };
 
-        let e3: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "class": ["person"],
-                "uuid": ["7b23c99d-c06b-4a9a-a958-3afa56383e1d"],
-                "uidnumber": ["1002"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e3 = unsafe { entry_init!(
+            ("class", CLASS_PERSON.clone()),
+            ("uuid", Value::Uuid(uuid::uuid!("7b23c99d-c06b-4a9a-a958-3afa56383e1d"))),
+            ("gidnumber", Value::Uint32(1002))
+        ).into_sealed_new() };
 
-        let e4: Entry<EntrySealed, EntryNew> = unsafe {
-            Entry::unsafe_from_entry_str(
-                r#"{
-            "attrs": {
-                "class": ["group"],
-                "uuid": ["21d816b5-1f6a-4696-b7c1-6ed06d22ed81"],
-                "uidnumber": ["1000"]
-            }
-        }"#,
-            )
-            .into_sealed_new()
-        };
+        let e4 = unsafe { entry_init!(
+            ("class", CLASS_GROUP.clone()),
+            ("uuid", Value::Uuid(uuid::uuid!("21d816b5-1f6a-4696-b7c1-6ed06d22ed81"))),
+            ("gidnumber", Value::Uint32(1000))
+        ).into_sealed_new() };
 
         let f_t1a = unsafe {
             filter_resolved!(f_and!([
-                f_eq("class", PartialValue::new_class("person")),
+                f_eq("class", PVCLASS_PERSON.clone()),
                 f_or!([
-                    f_eq("uidnumber", PartialValue::new_iutf8("1001")),
-                    f_eq("uidnumber", PartialValue::new_iutf8("1000"))
+                    f_eq("gidnumber", PartialValue::Uint32(1001)),
+                    f_eq("gidnumber", PartialValue::Uint32(1000))
                 ])
             ]))
         };
@@ -1809,33 +1753,36 @@ mod tests {
         let time_p3 = time_p2 + Duration::from_secs(CHANGELOG_MAX_AGE * 2);
 
         let mut server_txn = server.write(time_p1).await;
-        let e1: Entry<EntryInit, EntryNew> = Entry::unsafe_from_entry_str(
-            r#"{
-                "attrs": {
-                    "class": ["object", "person", "account"],
-                    "name": ["testperson1"],
-                    "uuid": ["cc8e95b4-c24f-4d68-ba54-8bed76f63930"],
-                    "description": ["testperson"],
-                    "displayname": ["testperson1"]
-                }
-            }"#,
+
+        let e1 = entry_init!(
+            ("class", CLASS_OBJECT.clone()),
+            ("class", CLASS_PERSON.clone()),
+            ("class", CLASS_ACCOUNT.clone()),
+            ("name", Value::new_iname("testperson1")),
+            ("uuid", Value::Uuid(uuid::uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))),
+            (
+                "description",
+                Value::new_utf8s("testperson1")
+            ),
+            ("displayname", Value::new_utf8s("testperson1"))
         );
-        let e2: Entry<EntryInit, EntryNew> = Entry::unsafe_from_entry_str(
-            r#"{
-                "attrs": {
-                    "class": ["object", "person"],
-                    "name": ["testperson2"],
-                    "uuid": ["a67c0c71-0b35-4218-a6b0-22d23d131d27"],
-                    "description": ["testperson"],
-                    "displayname": ["testperson2"]
-                }
-            }"#,
+
+        let e2 = entry_init!(
+            ("class", CLASS_OBJECT.clone()),
+            ("class", CLASS_PERSON.clone()),
+            ("name", Value::new_iname("testperson2")),
+            ("uuid", Value::Uuid(uuid::uuid!("a67c0c71-0b35-4218-a6b0-22d23d131d27"))),
+            (
+                "description",
+                Value::new_utf8s("testperson2")
+            ),
+            ("displayname", Value::new_utf8s("testperson2"))
         );
 
         // We need to add these and then push through the state machine.
         let e_ts = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
+            ("class", CLASS_OBJECT.clone()),
+            ("class", CLASS_PERSON.clone()),
             ("name", Value::new_iname("testperson3")),
             (
                 "uuid",
