+++ kanidmd/lib/src/server/migrations.rs
@@ -122,6 +122,69 @@ impl QueryServer {
 }
 
 impl<'a> QueryServerWriteTransaction<'a> {
+    #[instrument(level = "debug", skip_all)]
+    pub fn internal_migrate_or_create_str(&mut self, e_str: &str) -> Result<(), OperationError> {
+        let res = Entry::from_proto_entry_str(e_str, self)
+            /*
+            .and_then(|e: Entry<EntryInvalid, EntryNew>| {
+                let schema = self.get_schema();
+                e.validate(schema).map_err(OperationError::SchemaViolation)
+            })
+            */
+            .and_then(|e: Entry<EntryInit, EntryNew>| self.internal_migrate_or_create(e));
+        trace!(?res);
+        debug_assert!(res.is_ok());
+        res
+    }
+
+    pub fn internal_migrate_or_create(
+        &mut self,
+        e: Entry<EntryInit, EntryNew>,
+    ) -> Result<(), OperationError> {
+        // if the thing exists, ensure the set of attributes on
+        // Entry A match and are present (but don't delete multivalue, or extended
+        // attributes in the situation.
+        // If not exist, create from Entry B
+        //
+        // This will extra classes an attributes alone!
+        //
+        // NOTE: gen modlist IS schema aware and will handle multivalue
+        // correctly!
+        trace!("internal_migrate_or_create operating on {:?}", e.get_uuid());
+
+        let filt = match e.filter_from_attrs(&[AttrString::from("uuid")]) {
+            Some(f) => f,
+            None => return Err(OperationError::FilterGeneration),
+        };
+
+        trace!("internal_migrate_or_create search {:?}", filt);
+
+        let results = self.internal_search(filt.clone())?;
+
+        if results.is_empty() {
+            // It does not exist. Create it.
+            self.internal_create(vec![e])
+        } else if results.len() == 1 {
+            // If the thing is subset, pass
+            match e.gen_modlist_assert(&self.schema) {
+                Ok(modlist) => {
+                    // Apply to &results[0]
+                    trace!("Generated modlist -> {:?}", modlist);
+                    self.internal_modify(&filt, &modlist)
+                }
+                Err(e) => Err(OperationError::SchemaViolation(e)),
+            }
+        } else {
+            admin_error!(
+                "Invalid Result Set - Expected One Entry for {:?} - {:?}",
+                filt,
+                results
+            );
+            Err(OperationError::InvalidDbState)
+        }
+    }
+
+
     /// Migrate 2 to 3 changes the name, domain_name types from iutf8 to iname.
     #[instrument(level = "debug", skip_all)]
     pub fn migrate_2_to_3(&mut self) -> Result<(), OperationError> {
@@ -343,6 +406,416 @@ impl<'a> QueryServerWriteTransaction<'a> {
                 e
             })
         // Complete
+    }
+
+    #[instrument(level = "info", skip_all)]
+    pub fn initialise_schema_core(&mut self) -> Result<(), OperationError> {
+        admin_debug!("initialise_schema_core -> start ...");
+        // Load in all the "core" schema, that we already have in "memory".
+        let entries = self.schema.to_entries();
+
+        // admin_debug!("Dumping schemas: {:?}", entries);
+
+        // internal_migrate_or_create.
+        let r: Result<_, _> = entries.into_iter().try_for_each(|e| {
+            trace!(?e, "init schema entry");
+            self.internal_migrate_or_create(e)
+        });
+        if r.is_ok() {
+            admin_debug!("initialise_schema_core -> Ok!");
+        } else {
+            admin_error!(?r, "initialise_schema_core -> Error");
+        }
+        // why do we have error handling if it's always supposed to be `Ok`?
+        debug_assert!(r.is_ok());
+        r
+    }
+
+    #[instrument(level = "info", skip_all)]
+    pub fn initialise_schema_idm(&mut self) -> Result<(), OperationError> {
+        admin_debug!("initialise_schema_idm -> start ...");
+        // List of IDM schemas to init.
+        let idm_schema: Vec<&str> = vec![
+            JSON_SCHEMA_ATTR_DISPLAYNAME,
+            JSON_SCHEMA_ATTR_LEGALNAME,
+            JSON_SCHEMA_ATTR_MAIL,
+            JSON_SCHEMA_ATTR_SSH_PUBLICKEY,
+            JSON_SCHEMA_ATTR_PRIMARY_CREDENTIAL,
+            JSON_SCHEMA_ATTR_RADIUS_SECRET,
+            JSON_SCHEMA_ATTR_DOMAIN_NAME,
+            JSON_SCHEMA_ATTR_DOMAIN_DISPLAY_NAME,
+            JSON_SCHEMA_ATTR_DOMAIN_UUID,
+            JSON_SCHEMA_ATTR_DOMAIN_SSID,
+            JSON_SCHEMA_ATTR_DOMAIN_TOKEN_KEY,
+            JSON_SCHEMA_ATTR_FERNET_PRIVATE_KEY_STR,
+            JSON_SCHEMA_ATTR_GIDNUMBER,
+            JSON_SCHEMA_ATTR_BADLIST_PASSWORD,
+            JSON_SCHEMA_ATTR_LOGINSHELL,
+            JSON_SCHEMA_ATTR_UNIX_PASSWORD,
+            JSON_SCHEMA_ATTR_ACCOUNT_EXPIRE,
+            JSON_SCHEMA_ATTR_ACCOUNT_VALID_FROM,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_NAME,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_ORIGIN,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_SCOPE_MAP,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_IMPLICIT_SCOPES,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_BASIC_SECRET,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_TOKEN_KEY,
+            JSON_SCHEMA_ATTR_ES256_PRIVATE_KEY_DER,
+            JSON_SCHEMA_ATTR_OAUTH2_ALLOW_INSECURE_CLIENT_DISABLE_PKCE,
+            JSON_SCHEMA_ATTR_OAUTH2_JWT_LEGACY_CRYPTO_ENABLE,
+            JSON_SCHEMA_ATTR_RS256_PRIVATE_KEY_DER,
+            JSON_SCHEMA_ATTR_CREDENTIAL_UPDATE_INTENT_TOKEN,
+            JSON_SCHEMA_ATTR_OAUTH2_CONSENT_SCOPE_MAP,
+            JSON_SCHEMA_ATTR_PASSKEYS,
+            JSON_SCHEMA_ATTR_DEVICEKEYS,
+            JSON_SCHEMA_ATTR_DYNGROUP_FILTER,
+            JSON_SCHEMA_ATTR_JWS_ES256_PRIVATE_KEY,
+            JSON_SCHEMA_ATTR_API_TOKEN_SESSION,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_SUP_SCOPE_MAP,
+            JSON_SCHEMA_ATTR_USER_AUTH_TOKEN_SESSION,
+            JSON_SCHEMA_ATTR_OAUTH2_SESSION,
+            JSON_SCHEMA_ATTR_NSUNIQUEID,
+            JSON_SCHEMA_ATTR_OAUTH2_PREFER_SHORT_USERNAME,
+            JSON_SCHEMA_ATTR_SYNC_TOKEN_SESSION,
+            JSON_SCHEMA_ATTR_SYNC_COOKIE,
+            JSON_SCHEMA_ATTR_GRANT_UI_HINT,
+            JSON_SCHEMA_ATTR_OAUTH2_RS_ORIGIN_LANDING,
+            JSON_SCHEMA_CLASS_PERSON,
+            JSON_SCHEMA_CLASS_ORGPERSON,
+            JSON_SCHEMA_CLASS_GROUP,
+            JSON_SCHEMA_CLASS_DYNGROUP,
+            JSON_SCHEMA_CLASS_ACCOUNT,
+            JSON_SCHEMA_CLASS_SERVICE_ACCOUNT,
+            JSON_SCHEMA_CLASS_DOMAIN_INFO,
+            JSON_SCHEMA_CLASS_POSIXACCOUNT,
+            JSON_SCHEMA_CLASS_POSIXGROUP,
+            JSON_SCHEMA_CLASS_SYSTEM_CONFIG,
+            JSON_SCHEMA_CLASS_OAUTH2_RS,
+            JSON_SCHEMA_CLASS_OAUTH2_RS_BASIC,
+            JSON_SCHEMA_CLASS_SYNC_ACCOUNT,
+        ];
+
+        let r = idm_schema
+            .iter()
+            // Each item individually logs it's result
+            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
+
+        if r.is_ok() {
+            admin_debug!("initialise_schema_idm -> Ok!");
+        } else {
+            admin_error!(res = ?r, "initialise_schema_idm -> Error");
+        }
+        debug_assert!(r.is_ok()); // why return a result if we assert it's `Ok`?
+
+        r
+    }
+
+    // This function is idempotent
+    #[instrument(level = "info", skip_all)]
+    pub fn initialise_idm(&mut self) -> Result<(), OperationError> {
+        // First, check the system_info object. This stores some server information
+        // and details. It's a pretty const thing. Also check anonymous, important to many
+        // concepts.
+        let res = self
+            .internal_migrate_or_create_str(JSON_SYSTEM_INFO_V1)
+            .and_then(|_| self.internal_migrate_or_create_str(JSON_DOMAIN_INFO_V1))
+            .and_then(|_| self.internal_migrate_or_create_str(JSON_SYSTEM_CONFIG_V1));
+        if res.is_err() {
+            admin_error!("initialise_idm p1 -> result {:?}", res);
+        }
+        debug_assert!(res.is_ok());
+        res?;
+
+        // The domain info now exists, we should be able to do these migrations as they will
+        // cause SPN regenerations to occur
+
+        // Check the admin object exists (migrations).
+        // Create the default idm_admin group.
+        let admin_entries = [
+            JSON_ANONYMOUS_V1,
+            JSON_ADMIN_V1,
+            JSON_IDM_ADMIN_V1,
+            JSON_IDM_ADMINS_V1,
+            JSON_SYSTEM_ADMINS_V1,
+        ];
+        let res: Result<(), _> = admin_entries
+            .iter()
+            // Each item individually logs it's result
+            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
+        if res.is_err() {
+            admin_error!("initialise_idm p2 -> result {:?}", res);
+        }
+        debug_assert!(res.is_ok());
+        res?;
+
+        // Create any system default schema entries.
+
+        // Create any system default access profile entries.
+        let idm_entries = [
+            // Builtin dyn groups,
+            JSON_IDM_ALL_PERSONS,
+            JSON_IDM_ALL_ACCOUNTS,
+            // Builtin groups
+            JSON_IDM_PEOPLE_MANAGE_PRIV_V1,
+            JSON_IDM_PEOPLE_ACCOUNT_PASSWORD_IMPORT_PRIV_V1,
+            JSON_IDM_PEOPLE_EXTEND_PRIV_V1,
+            JSON_IDM_PEOPLE_SELF_WRITE_MAIL_PRIV_V1,
+            JSON_IDM_PEOPLE_WRITE_PRIV_V1,
+            JSON_IDM_PEOPLE_READ_PRIV_V1,
+            JSON_IDM_HP_PEOPLE_EXTEND_PRIV_V1,
+            JSON_IDM_HP_PEOPLE_WRITE_PRIV_V1,
+            JSON_IDM_HP_PEOPLE_READ_PRIV_V1,
+            JSON_IDM_GROUP_MANAGE_PRIV_V1,
+            JSON_IDM_GROUP_WRITE_PRIV_V1,
+            JSON_IDM_GROUP_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_ACCOUNT_MANAGE_PRIV_V1,
+            JSON_IDM_ACCOUNT_WRITE_PRIV_V1,
+            JSON_IDM_ACCOUNT_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_ACCOUNT_READ_PRIV_V1,
+            JSON_IDM_RADIUS_SECRET_WRITE_PRIV_V1,
+            JSON_IDM_RADIUS_SECRET_READ_PRIV_V1,
+            JSON_IDM_RADIUS_SERVERS_V1,
+            // Write deps on read, so write must be added first.
+            JSON_IDM_HP_ACCOUNT_MANAGE_PRIV_V1,
+            JSON_IDM_HP_ACCOUNT_WRITE_PRIV_V1,
+            JSON_IDM_HP_ACCOUNT_READ_PRIV_V1,
+            JSON_IDM_HP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_SCHEMA_MANAGE_PRIV_V1,
+            JSON_IDM_HP_GROUP_MANAGE_PRIV_V1,
+            JSON_IDM_HP_GROUP_WRITE_PRIV_V1,
+            JSON_IDM_HP_GROUP_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_ACP_MANAGE_PRIV_V1,
+            JSON_DOMAIN_ADMINS,
+            JSON_IDM_HP_OAUTH2_MANAGE_PRIV_V1,
+            JSON_IDM_HP_SERVICE_ACCOUNT_INTO_PERSON_MIGRATE_PRIV,
+            JSON_IDM_HP_SYNC_ACCOUNT_MANAGE_PRIV,
+            // All members must exist before we write HP
+            JSON_IDM_HIGH_PRIVILEGE_V1,
+            // Built in access controls.
+            JSON_IDM_ADMINS_ACP_RECYCLE_SEARCH_V1,
+            JSON_IDM_ADMINS_ACP_REVIVE_V1,
+            // JSON_IDM_ADMINS_ACP_MANAGE_V1,
+            JSON_IDM_ALL_ACP_READ_V1,
+            JSON_IDM_SELF_ACP_READ_V1,
+            JSON_IDM_SELF_ACP_WRITE_V1,
+            JSON_IDM_PEOPLE_SELF_ACP_WRITE_MAIL_PRIV_V1,
+            JSON_IDM_ACP_PEOPLE_READ_PRIV_V1,
+            JSON_IDM_ACP_PEOPLE_WRITE_PRIV_V1,
+            JSON_IDM_ACP_PEOPLE_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_GROUP_WRITE_PRIV_V1,
+            JSON_IDM_ACP_GROUP_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_ACCOUNT_READ_PRIV_V1,
+            JSON_IDM_ACP_ACCOUNT_WRITE_PRIV_V1,
+            JSON_IDM_ACP_ACCOUNT_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_RADIUS_SERVERS_V1,
+            JSON_IDM_ACP_HP_ACCOUNT_READ_PRIV_V1,
+            JSON_IDM_ACP_HP_ACCOUNT_WRITE_PRIV_V1,
+            JSON_IDM_ACP_HP_ACCOUNT_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_HP_GROUP_WRITE_PRIV_V1,
+            JSON_IDM_ACP_HP_GROUP_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_SCHEMA_WRITE_ATTRS_PRIV_V1,
+            JSON_IDM_ACP_SCHEMA_WRITE_CLASSES_PRIV_V1,
+            JSON_IDM_ACP_ACP_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_DOMAIN_ADMIN_PRIV_V1,
+            JSON_IDM_ACP_SYSTEM_CONFIG_PRIV_V1,
+            JSON_IDM_ACP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_ACP_GROUP_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_ACP_PEOPLE_ACCOUNT_PASSWORD_IMPORT_PRIV_V1,
+            JSON_IDM_ACP_PEOPLE_EXTEND_PRIV_V1,
+            JSON_IDM_ACP_HP_PEOPLE_READ_PRIV_V1,
+            JSON_IDM_ACP_HP_PEOPLE_WRITE_PRIV_V1,
+            JSON_IDM_ACP_HP_PEOPLE_EXTEND_PRIV_V1,
+            JSON_IDM_HP_ACP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_HP_ACP_GROUP_UNIX_EXTEND_PRIV_V1,
+            JSON_IDM_HP_ACP_OAUTH2_MANAGE_PRIV_V1,
+            JSON_IDM_ACP_RADIUS_SECRET_READ_PRIV_V1,
+            JSON_IDM_ACP_RADIUS_SECRET_WRITE_PRIV_V1,
+            JSON_IDM_HP_ACP_SERVICE_ACCOUNT_INTO_PERSON_MIGRATE_V1,
+            JSON_IDM_ACP_OAUTH2_READ_PRIV_V1,
+            JSON_IDM_HP_ACP_SYNC_ACCOUNT_MANAGE_PRIV_V1,
+        ];
+
+        let res: Result<(), _> = idm_entries
+            .iter()
+            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
+        if res.is_ok() {
+            admin_debug!("initialise_idm -> result Ok!");
+        } else {
+            admin_error!(?res, "initialise_idm p3 -> result");
+        }
+        debug_assert!(res.is_ok());
+        res?;
+
+        let idm_entries = [
+            E_IDM_UI_ENABLE_EXPERIMENTAL_FEATURES.clone(),
+            E_IDM_ACCOUNT_MAIL_READ_PRIV.clone(),
+            E_IDM_ACP_ACCOUNT_MAIL_READ_PRIV_V1.clone(),
+        ];
+
+        let res: Result<(), _> = idm_entries
+            .into_iter()
+            .try_for_each(|entry| self.internal_migrate_or_create(entry));
+        if res.is_ok() {
+            admin_debug!("initialise_idm -> result Ok!");
+        } else {
+            admin_error!(?res, "initialise_idm p3 -> result");
+        }
+        debug_assert!(res.is_ok());
+        res?;
+
+        self.changed_schema.set(true);
+        self.changed_acp.set(true);
+
+        Ok(())
+    }
+
+
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::prelude::*;
+
+    #[qs_test]
+    async fn test_init_idempotent_schema_core(server: &QueryServer) {
+        {
+            // Setup and abort.
+            let mut server_txn = server.write(duration_from_epoch_now()).await;
+            assert!(server_txn.initialise_schema_core().is_ok());
+        }
+        {
+            let mut server_txn = server.write(duration_from_epoch_now()).await;
+            assert!(server_txn.initialise_schema_core().is_ok());
+            assert!(server_txn.initialise_schema_core().is_ok());
+            assert!(server_txn.commit().is_ok());
+        }
+        {
+            // Now do it again in a new txn, but abort
+            let mut server_txn = server.write(duration_from_epoch_now()).await;
+            assert!(server_txn.initialise_schema_core().is_ok());
+        }
+        {
+            // Now do it again in a new txn.
+            let mut server_txn = server.write(duration_from_epoch_now()).await;
+            assert!(server_txn.initialise_schema_core().is_ok());
+            assert!(server_txn.commit().is_ok());
+        }
+    }
+
+    #[qs_test_no_init]
+    async fn test_qs_upgrade_entry_attrs(server: &QueryServer) {
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        assert!(server_txn.upgrade_reindex(SYSTEM_INDEX_VERSION).is_ok());
+        assert!(server_txn.commit().is_ok());
+
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        server_txn.initialise_schema_core().unwrap();
+        server_txn.initialise_schema_idm().unwrap();
+        assert!(server_txn.commit().is_ok());
+
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        assert!(server_txn.upgrade_reindex(SYSTEM_INDEX_VERSION + 1).is_ok());
+        assert!(server_txn.commit().is_ok());
+
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        assert!(server_txn
+            .internal_migrate_or_create_str(JSON_SYSTEM_INFO_V1)
+            .is_ok());
+        assert!(server_txn
+            .internal_migrate_or_create_str(JSON_DOMAIN_INFO_V1)
+            .is_ok());
+        assert!(server_txn
+            .internal_migrate_or_create_str(JSON_SYSTEM_CONFIG_V1)
+            .is_ok());
+        assert!(server_txn.commit().is_ok());
+
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        // ++ Mod the schema to set name to the old string type
+        let me_syn = unsafe {
+            ModifyEvent::new_internal_invalid(
+                filter!(f_or!([
+                    f_eq("attributename", PartialValue::new_iutf8("name")),
+                    f_eq("attributename", PartialValue::new_iutf8("domain_name")),
+                ])),
+                ModifyList::new_purge_and_set(
+                    "syntax",
+                    Value::new_syntaxs("UTF8STRING_INSENSITIVE").unwrap(),
+                ),
+            )
+        };
+        assert!(server_txn.modify(&me_syn).is_ok());
+        assert!(server_txn.commit().is_ok());
+
+        let mut server_txn = server.write(duration_from_epoch_now()).await;
+        // ++ Mod domain name and name to be the old type.
+        let me_dn = unsafe {
+            ModifyEvent::new_internal_invalid(
+                filter!(f_eq("uuid", PartialValue::Uuid(UUID_DOMAIN_INFO))),
+                ModifyList::new_list(vec![
+                    Modify::Purged(AttrString::from("name")),
+                    Modify::Purged(AttrString::from("domain_name")),
+                    Modify::Present(AttrString::from("name"), Value::new_iutf8("domain_local")),
+                    Modify::Present(
+                        AttrString::from("domain_name"),
+                        Value::new_iutf8("example.com"),
+                    ),
+                ]),
+            )
+        };
+        assert!(server_txn.modify(&me_dn).is_ok());
+
+        // Now, both the types are invalid.
+
+        // WARNING! We can't commit here because this triggers domain_reload which will fail
+        // due to incorrect syntax of the domain name! Run the migration in the same txn!
+        // Trigger a schema reload.
+        assert!(server_txn.reload_schema().is_ok());
+
+        // We can't just re-run the migrate here because name takes it's definition from
+        // in memory, and we can't re-run the initial memory gen. So we just fix it to match
+        // what the migrate "would do".
+        let me_syn = unsafe {
+            ModifyEvent::new_internal_invalid(
+                filter!(f_or!([
+                    f_eq("attributename", PartialValue::new_iutf8("name")),
+                    f_eq("attributename", PartialValue::new_iutf8("domain_name")),
+                ])),
+                ModifyList::new_purge_and_set(
+                    "syntax",
+                    Value::new_syntaxs("UTF8STRING_INAME").unwrap(),
+                ),
+            )
+        };
+        assert!(server_txn.modify(&me_syn).is_ok());
+
+        // WARNING! We can't commit here because this triggers domain_reload which will fail
+        // due to incorrect syntax of the domain name! Run the migration in the same txn!
+        // Trigger a schema reload.
+        assert!(server_txn.reload_schema().is_ok());
+
+        // ++ Run the upgrade for X to Y
+        assert!(server_txn.migrate_2_to_3().is_ok());
+
+        assert!(server_txn.commit().is_ok());
+
+        // Assert that it migrated and worked as expected.
+        let server_txn = server.write(duration_from_epoch_now()).await;
+        let domain = server_txn
+            .internal_search_uuid(UUID_DOMAIN_INFO)
+            .expect("failed");
+        // ++ assert all names are iname
+        assert!(
+            domain.get_ava_set("name").expect("no name?").syntax() == SyntaxType::Utf8StringIname
+        );
+        // ++ assert all domain/domain_name are iname
+        assert!(
+            domain
+                .get_ava_set("domain_name")
+                .expect("no domain_name?")
+                .syntax()
+                == SyntaxType::Utf8StringIname
+        );
+        assert!(server_txn.commit().is_ok());
     }
 }
