+++ kanidmd/lib/src/server/mod.rs
@@ -6,7 +6,8 @@
 use std::cell::Cell;
 use std::str::FromStr;
 use std::sync::Arc;
-use std::time::Duration;
+
+use crate::prelude::*;
 
 use concread::arcache::{ARCache, ARCacheBuilder, ARCacheReadTxn};
 use concread::cowcell::*;
@@ -22,15 +23,10 @@ use crate::access::{
 };
 use crate::be::{Backend, BackendReadTransaction, BackendTransaction, BackendWriteTransaction};
 // We use so many, we just import them all ...
-use crate::event::{
-    CreateEvent, DeleteEvent, ExistsEvent, ModifyEvent, SearchEvent,
-};
 use crate::filter::{Filter, FilterInvalid, FilterValid, FilterValidResolved};
 use crate::identity::IdentityId;
-use crate::modify::{ModifyInvalid, ModifyList, ModifyValid};
 use crate::plugins::dyngroup::{DynGroup, DynGroupCache};
 use crate::plugins::Plugins;
-use crate::prelude::*;
 use crate::repl::cid::Cid;
 use crate::schema::{
     Schema, SchemaAttribute, SchemaClass, SchemaReadTransaction, SchemaTransaction,
@@ -121,9 +117,6 @@ pub struct QueryServerWriteTransaction<'a> {
     dyngroup_cache: Cell<CowCellWriteTxn<'a, DynGroupCache>>,
 }
 
-// This is the core of the server. It implements all
-// the search and modify actions, applies access controls
-// and get's everything ready to push back to the fe code
 /// The `QueryServerTransaction` trait provides a set of common read only operations to be
 /// shared between [`QueryServerReadTransaction`] and [`QueryServerWriteTransaction`]s.
 ///
@@ -1035,534 +1028,6 @@ impl<'a> QueryServerWriteTransaction<'a> {
         }
     }
 
-    // These are where searches and other actions are actually implemented. This
-    // is the "internal" version, where we define the event as being internal
-    // only, allowing certain plugin by passes etc.
-
-    pub fn internal_create(
-        &mut self,
-        entries: Vec<Entry<EntryInit, EntryNew>>,
-    ) -> Result<(), OperationError> {
-        let ce = CreateEvent::new_internal(entries);
-        self.create(&ce)
-    }
-
-    pub fn internal_delete(
-        &mut self,
-        filter: &Filter<FilterInvalid>,
-    ) -> Result<(), OperationError> {
-        let f_valid = filter
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let de = DeleteEvent::new_internal(f_valid);
-        self.delete(&de)
-    }
-
-    pub fn internal_delete_uuid(&mut self, target_uuid: Uuid) -> Result<(), OperationError> {
-        let filter = filter!(f_eq("uuid", PartialValue::Uuid(target_uuid)));
-        let f_valid = filter
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let de = DeleteEvent::new_internal(f_valid);
-        self.delete(&de)
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    pub fn internal_modify(
-        &mut self,
-        filter: &Filter<FilterInvalid>,
-        modlist: &ModifyList<ModifyInvalid>,
-    ) -> Result<(), OperationError> {
-        let f_valid = filter
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let m_valid = modlist
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let me = ModifyEvent::new_internal(f_valid, m_valid);
-        self.modify(&me)
-    }
-
-    pub fn internal_modify_uuid(
-        &mut self,
-        target_uuid: Uuid,
-        modlist: &ModifyList<ModifyInvalid>,
-    ) -> Result<(), OperationError> {
-        let filter = filter!(f_eq("uuid", PartialValue::Uuid(target_uuid)));
-        let f_valid = filter
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let m_valid = modlist
-            .validate(self.get_schema())
-            .map_err(OperationError::SchemaViolation)?;
-        let me = ModifyEvent::new_internal(f_valid, m_valid);
-        self.modify(&me)
-    }
-
-    pub fn impersonate_modify_valid(
-        &mut self,
-        f_valid: Filter<FilterValid>,
-        f_intent_valid: Filter<FilterValid>,
-        m_valid: ModifyList<ModifyValid>,
-        event: &Identity,
-    ) -> Result<(), OperationError> {
-        let me = ModifyEvent::new_impersonate(event, f_valid, f_intent_valid, m_valid);
-        self.modify(&me)
-    }
-
-    pub fn impersonate_modify(
-        &mut self,
-        filter: &Filter<FilterInvalid>,
-        filter_intent: &Filter<FilterInvalid>,
-        modlist: &ModifyList<ModifyInvalid>,
-        event: &Identity,
-    ) -> Result<(), OperationError> {
-        let f_valid = filter.validate(self.get_schema()).map_err(|e| {
-            admin_error!("filter Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        let f_intent_valid = filter_intent.validate(self.get_schema()).map_err(|e| {
-            admin_error!("f_intent Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        let m_valid = modlist.validate(self.get_schema()).map_err(|e| {
-            admin_error!("modlist Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        self.impersonate_modify_valid(f_valid, f_intent_valid, m_valid, event)
-    }
-
-    pub fn impersonate_modify_gen_event(
-        &mut self,
-        filter: &Filter<FilterInvalid>,
-        filter_intent: &Filter<FilterInvalid>,
-        modlist: &ModifyList<ModifyInvalid>,
-        event: &Identity,
-    ) -> Result<ModifyEvent, OperationError> {
-        let f_valid = filter.validate(self.get_schema()).map_err(|e| {
-            admin_error!("filter Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        let f_intent_valid = filter_intent.validate(self.get_schema()).map_err(|e| {
-            admin_error!("f_intent Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        let m_valid = modlist.validate(self.get_schema()).map_err(|e| {
-            admin_error!("modlist Schema Invalid {:?}", e);
-            OperationError::SchemaViolation(e)
-        })?;
-        Ok(ModifyEvent::new_impersonate(
-            event,
-            f_valid,
-            f_intent_valid,
-            m_valid,
-        ))
-    }
-
-    // internal server operation types.
-    // These just wrap the fn create/search etc, but they allow
-    // creating the needed create event with the correct internal flags
-    // and markers. They act as though they have the highest level privilege
-    // IE there are no access control checks.
-
-    /*
-    pub fn internal_exists_or_create(
-        &self,
-        _e: Entry<EntryValid, EntryNew>,
-    ) -> Result<(), OperationError> {
-        // If the thing exists, stop.
-        // if not, create from Entry.
-        unimplemented!()
-    }
-    */
-
-    #[instrument(level = "debug", skip_all)]
-    pub fn internal_migrate_or_create_str(&mut self, e_str: &str) -> Result<(), OperationError> {
-        let res = Entry::from_proto_entry_str(e_str, self)
-            /*
-            .and_then(|e: Entry<EntryInvalid, EntryNew>| {
-                let schema = self.get_schema();
-                e.validate(schema).map_err(OperationError::SchemaViolation)
-            })
-            */
-            .and_then(|e: Entry<EntryInit, EntryNew>| self.internal_migrate_or_create(e));
-        trace!(?res);
-        debug_assert!(res.is_ok());
-        res
-    }
-
-    pub fn internal_migrate_or_create(
-        &mut self,
-        e: Entry<EntryInit, EntryNew>,
-    ) -> Result<(), OperationError> {
-        // if the thing exists, ensure the set of attributes on
-        // Entry A match and are present (but don't delete multivalue, or extended
-        // attributes in the situation.
-        // If not exist, create from Entry B
-        //
-        // This will extra classes an attributes alone!
-        //
-        // NOTE: gen modlist IS schema aware and will handle multivalue
-        // correctly!
-        trace!("internal_migrate_or_create operating on {:?}", e.get_uuid());
-
-        let filt = match e.filter_from_attrs(&[AttrString::from("uuid")]) {
-            Some(f) => f,
-            None => return Err(OperationError::FilterGeneration),
-        };
-
-        trace!("internal_migrate_or_create search {:?}", filt);
-
-        let results = self.internal_search(filt.clone())?;
-
-        if results.is_empty() {
-            // It does not exist. Create it.
-            self.internal_create(vec![e])
-        } else if results.len() == 1 {
-            // If the thing is subset, pass
-            match e.gen_modlist_assert(&self.schema) {
-                Ok(modlist) => {
-                    // Apply to &results[0]
-                    trace!("Generated modlist -> {:?}", modlist);
-                    self.internal_modify(&filt, &modlist)
-                }
-                Err(e) => Err(OperationError::SchemaViolation(e)),
-            }
-        } else {
-            admin_error!(
-                "Invalid Result Set - Expected One Entry for {:?} - {:?}",
-                filt,
-                results
-            );
-            Err(OperationError::InvalidDbState)
-        }
-    }
-
-    /*
-    pub fn internal_assert_or_create_str(
-        &mut self,
-        e_str: &str,
-    ) -> Result<(), OperationError> {
-        let res = audit_segment!( || Entry::from_proto_entry_str( e_str, self)
-            .and_then(
-                |e: Entry<EntryInit, EntryNew>| self.internal_assert_or_create( e)
-            ));
-        ltrace!( "internal_assert_or_create_str -> result {:?}", res);
-        debug_assert!(res.is_ok());
-        res
-    }
-
-    // Should this take a be_txn?
-    pub fn internal_assert_or_create(
-        &mut self,
-        e: Entry<EntryInit, EntryNew>,
-    ) -> Result<(), OperationError> {
-        // If exists, ensure the object is exactly as provided
-        // else, if not exists, create it. IE no extra or excess
-        // attributes and classes.
-
-        ltrace!(
-
-            "internal_assert_or_create operating on {:?}",
-            e.get_uuid()
-        );
-
-        // Create a filter from the entry for assertion.
-        let filt = match e.filter_from_attrs(&[String::from("uuid")]) {
-            Some(f) => f,
-            None => return Err(OperationError::FilterGeneration),
-        };
-
-        // Does it exist? we use search here, not exists, so that if the entry does exist
-        // we can compare it is identical, which avoids a delete/create cycle that would
-        // trigger csn/repl each time we start up.
-        match self.internal_search( filt.clone()) {
-            Ok(results) => {
-                if results.is_empty() {
-                    // It does not exist. Create it.
-                    self.internal_create( vec![e])
-                } else if results.len() == 1 {
-                    // it exists. To guarantee content exactly as is, we compare if it's identical.
-                    if !e.compare(&results[0]) {
-                        self.internal_delete( filt)
-                            .and_then(|_| self.internal_create( vec![e]))
-                    } else {
-                        // No action required
-                        Ok(())
-                    }
-                } else {
-                    Err(OperationError::InvalidDbState)
-                }
-            }
-            Err(er) => {
-                // An error occured. pass it back up.
-                Err(er)
-            }
-        }
-    }
-    */
-
-    #[instrument(level = "info", skip_all)]
-    pub fn initialise_schema_core(&mut self) -> Result<(), OperationError> {
-        admin_debug!("initialise_schema_core -> start ...");
-        // Load in all the "core" schema, that we already have in "memory".
-        let entries = self.schema.to_entries();
-
-        // admin_debug!("Dumping schemas: {:?}", entries);
-
-        // internal_migrate_or_create.
-        let r: Result<_, _> = entries.into_iter().try_for_each(|e| {
-            trace!(?e, "init schema entry");
-            self.internal_migrate_or_create(e)
-        });
-        if r.is_ok() {
-            admin_debug!("initialise_schema_core -> Ok!");
-        } else {
-            admin_error!(?r, "initialise_schema_core -> Error");
-        }
-        // why do we have error handling if it's always supposed to be `Ok`?
-        debug_assert!(r.is_ok());
-        r
-    }
-
-    #[instrument(level = "info", skip_all)]
-    pub fn initialise_schema_idm(&mut self) -> Result<(), OperationError> {
-        admin_debug!("initialise_schema_idm -> start ...");
-        // List of IDM schemas to init.
-        let idm_schema: Vec<&str> = vec![
-            JSON_SCHEMA_ATTR_DISPLAYNAME,
-            JSON_SCHEMA_ATTR_LEGALNAME,
-            JSON_SCHEMA_ATTR_MAIL,
-            JSON_SCHEMA_ATTR_SSH_PUBLICKEY,
-            JSON_SCHEMA_ATTR_PRIMARY_CREDENTIAL,
-            JSON_SCHEMA_ATTR_RADIUS_SECRET,
-            JSON_SCHEMA_ATTR_DOMAIN_NAME,
-            JSON_SCHEMA_ATTR_DOMAIN_DISPLAY_NAME,
-            JSON_SCHEMA_ATTR_DOMAIN_UUID,
-            JSON_SCHEMA_ATTR_DOMAIN_SSID,
-            JSON_SCHEMA_ATTR_DOMAIN_TOKEN_KEY,
-            JSON_SCHEMA_ATTR_FERNET_PRIVATE_KEY_STR,
-            JSON_SCHEMA_ATTR_GIDNUMBER,
-            JSON_SCHEMA_ATTR_BADLIST_PASSWORD,
-            JSON_SCHEMA_ATTR_LOGINSHELL,
-            JSON_SCHEMA_ATTR_UNIX_PASSWORD,
-            JSON_SCHEMA_ATTR_ACCOUNT_EXPIRE,
-            JSON_SCHEMA_ATTR_ACCOUNT_VALID_FROM,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_NAME,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_ORIGIN,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_SCOPE_MAP,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_IMPLICIT_SCOPES,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_BASIC_SECRET,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_TOKEN_KEY,
-            JSON_SCHEMA_ATTR_ES256_PRIVATE_KEY_DER,
-            JSON_SCHEMA_ATTR_OAUTH2_ALLOW_INSECURE_CLIENT_DISABLE_PKCE,
-            JSON_SCHEMA_ATTR_OAUTH2_JWT_LEGACY_CRYPTO_ENABLE,
-            JSON_SCHEMA_ATTR_RS256_PRIVATE_KEY_DER,
-            JSON_SCHEMA_ATTR_CREDENTIAL_UPDATE_INTENT_TOKEN,
-            JSON_SCHEMA_ATTR_OAUTH2_CONSENT_SCOPE_MAP,
-            JSON_SCHEMA_ATTR_PASSKEYS,
-            JSON_SCHEMA_ATTR_DEVICEKEYS,
-            JSON_SCHEMA_ATTR_DYNGROUP_FILTER,
-            JSON_SCHEMA_ATTR_JWS_ES256_PRIVATE_KEY,
-            JSON_SCHEMA_ATTR_API_TOKEN_SESSION,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_SUP_SCOPE_MAP,
-            JSON_SCHEMA_ATTR_USER_AUTH_TOKEN_SESSION,
-            JSON_SCHEMA_ATTR_OAUTH2_SESSION,
-            JSON_SCHEMA_ATTR_NSUNIQUEID,
-            JSON_SCHEMA_ATTR_OAUTH2_PREFER_SHORT_USERNAME,
-            JSON_SCHEMA_ATTR_SYNC_TOKEN_SESSION,
-            JSON_SCHEMA_ATTR_SYNC_COOKIE,
-            JSON_SCHEMA_ATTR_GRANT_UI_HINT,
-            JSON_SCHEMA_ATTR_OAUTH2_RS_ORIGIN_LANDING,
-            JSON_SCHEMA_CLASS_PERSON,
-            JSON_SCHEMA_CLASS_ORGPERSON,
-            JSON_SCHEMA_CLASS_GROUP,
-            JSON_SCHEMA_CLASS_DYNGROUP,
-            JSON_SCHEMA_CLASS_ACCOUNT,
-            JSON_SCHEMA_CLASS_SERVICE_ACCOUNT,
-            JSON_SCHEMA_CLASS_DOMAIN_INFO,
-            JSON_SCHEMA_CLASS_POSIXACCOUNT,
-            JSON_SCHEMA_CLASS_POSIXGROUP,
-            JSON_SCHEMA_CLASS_SYSTEM_CONFIG,
-            JSON_SCHEMA_CLASS_OAUTH2_RS,
-            JSON_SCHEMA_CLASS_OAUTH2_RS_BASIC,
-            JSON_SCHEMA_CLASS_SYNC_ACCOUNT,
-        ];
-
-        let r = idm_schema
-            .iter()
-            // Each item individually logs it's result
-            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
-
-        if r.is_ok() {
-            admin_debug!("initialise_schema_idm -> Ok!");
-        } else {
-            admin_error!(res = ?r, "initialise_schema_idm -> Error");
-        }
-        debug_assert!(r.is_ok()); // why return a result if we assert it's `Ok`?
-
-        r
-    }
-
-    // This function is idempotent
-    #[instrument(level = "info", skip_all)]
-    pub fn initialise_idm(&mut self) -> Result<(), OperationError> {
-        // First, check the system_info object. This stores some server information
-        // and details. It's a pretty const thing. Also check anonymous, important to many
-        // concepts.
-        let res = self
-            .internal_migrate_or_create_str(JSON_SYSTEM_INFO_V1)
-            .and_then(|_| self.internal_migrate_or_create_str(JSON_DOMAIN_INFO_V1))
-            .and_then(|_| self.internal_migrate_or_create_str(JSON_SYSTEM_CONFIG_V1));
-        if res.is_err() {
-            admin_error!("initialise_idm p1 -> result {:?}", res);
-        }
-        debug_assert!(res.is_ok());
-        res?;
-
-        // The domain info now exists, we should be able to do these migrations as they will
-        // cause SPN regenerations to occur
-
-        // Check the admin object exists (migrations).
-        // Create the default idm_admin group.
-        let admin_entries = [
-            JSON_ANONYMOUS_V1,
-            JSON_ADMIN_V1,
-            JSON_IDM_ADMIN_V1,
-            JSON_IDM_ADMINS_V1,
-            JSON_SYSTEM_ADMINS_V1,
-        ];
-        let res: Result<(), _> = admin_entries
-            .iter()
-            // Each item individually logs it's result
-            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
-        if res.is_err() {
-            admin_error!("initialise_idm p2 -> result {:?}", res);
-        }
-        debug_assert!(res.is_ok());
-        res?;
-
-        // Create any system default schema entries.
-
-        // Create any system default access profile entries.
-        let idm_entries = [
-            // Builtin dyn groups,
-            JSON_IDM_ALL_PERSONS,
-            JSON_IDM_ALL_ACCOUNTS,
-            // Builtin groups
-            JSON_IDM_PEOPLE_MANAGE_PRIV_V1,
-            JSON_IDM_PEOPLE_ACCOUNT_PASSWORD_IMPORT_PRIV_V1,
-            JSON_IDM_PEOPLE_EXTEND_PRIV_V1,
-            JSON_IDM_PEOPLE_SELF_WRITE_MAIL_PRIV_V1,
-            JSON_IDM_PEOPLE_WRITE_PRIV_V1,
-            JSON_IDM_PEOPLE_READ_PRIV_V1,
-            JSON_IDM_HP_PEOPLE_EXTEND_PRIV_V1,
-            JSON_IDM_HP_PEOPLE_WRITE_PRIV_V1,
-            JSON_IDM_HP_PEOPLE_READ_PRIV_V1,
-            JSON_IDM_GROUP_MANAGE_PRIV_V1,
-            JSON_IDM_GROUP_WRITE_PRIV_V1,
-            JSON_IDM_GROUP_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_ACCOUNT_MANAGE_PRIV_V1,
-            JSON_IDM_ACCOUNT_WRITE_PRIV_V1,
-            JSON_IDM_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_ACCOUNT_READ_PRIV_V1,
-            JSON_IDM_RADIUS_SECRET_WRITE_PRIV_V1,
-            JSON_IDM_RADIUS_SECRET_READ_PRIV_V1,
-            JSON_IDM_RADIUS_SERVERS_V1,
-            // Write deps on read, so write must be added first.
-            JSON_IDM_HP_ACCOUNT_MANAGE_PRIV_V1,
-            JSON_IDM_HP_ACCOUNT_WRITE_PRIV_V1,
-            JSON_IDM_HP_ACCOUNT_READ_PRIV_V1,
-            JSON_IDM_HP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_SCHEMA_MANAGE_PRIV_V1,
-            JSON_IDM_HP_GROUP_MANAGE_PRIV_V1,
-            JSON_IDM_HP_GROUP_WRITE_PRIV_V1,
-            JSON_IDM_HP_GROUP_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_ACP_MANAGE_PRIV_V1,
-            JSON_DOMAIN_ADMINS,
-            JSON_IDM_HP_OAUTH2_MANAGE_PRIV_V1,
-            JSON_IDM_HP_SERVICE_ACCOUNT_INTO_PERSON_MIGRATE_PRIV,
-            JSON_IDM_HP_SYNC_ACCOUNT_MANAGE_PRIV,
-            // All members must exist before we write HP
-            JSON_IDM_HIGH_PRIVILEGE_V1,
-            // Built in access controls.
-            JSON_IDM_ADMINS_ACP_RECYCLE_SEARCH_V1,
-            JSON_IDM_ADMINS_ACP_REVIVE_V1,
-            // JSON_IDM_ADMINS_ACP_MANAGE_V1,
-            JSON_IDM_ALL_ACP_READ_V1,
-            JSON_IDM_SELF_ACP_READ_V1,
-            JSON_IDM_SELF_ACP_WRITE_V1,
-            JSON_IDM_PEOPLE_SELF_ACP_WRITE_MAIL_PRIV_V1,
-            JSON_IDM_ACP_PEOPLE_READ_PRIV_V1,
-            JSON_IDM_ACP_PEOPLE_WRITE_PRIV_V1,
-            JSON_IDM_ACP_PEOPLE_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_GROUP_WRITE_PRIV_V1,
-            JSON_IDM_ACP_GROUP_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_ACCOUNT_READ_PRIV_V1,
-            JSON_IDM_ACP_ACCOUNT_WRITE_PRIV_V1,
-            JSON_IDM_ACP_ACCOUNT_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_RADIUS_SERVERS_V1,
-            JSON_IDM_ACP_HP_ACCOUNT_READ_PRIV_V1,
-            JSON_IDM_ACP_HP_ACCOUNT_WRITE_PRIV_V1,
-            JSON_IDM_ACP_HP_ACCOUNT_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_HP_GROUP_WRITE_PRIV_V1,
-            JSON_IDM_ACP_HP_GROUP_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_SCHEMA_WRITE_ATTRS_PRIV_V1,
-            JSON_IDM_ACP_SCHEMA_WRITE_CLASSES_PRIV_V1,
-            JSON_IDM_ACP_ACP_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_DOMAIN_ADMIN_PRIV_V1,
-            JSON_IDM_ACP_SYSTEM_CONFIG_PRIV_V1,
-            JSON_IDM_ACP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_ACP_GROUP_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_ACP_PEOPLE_ACCOUNT_PASSWORD_IMPORT_PRIV_V1,
-            JSON_IDM_ACP_PEOPLE_EXTEND_PRIV_V1,
-            JSON_IDM_ACP_HP_PEOPLE_READ_PRIV_V1,
-            JSON_IDM_ACP_HP_PEOPLE_WRITE_PRIV_V1,
-            JSON_IDM_ACP_HP_PEOPLE_EXTEND_PRIV_V1,
-            JSON_IDM_HP_ACP_ACCOUNT_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_HP_ACP_GROUP_UNIX_EXTEND_PRIV_V1,
-            JSON_IDM_HP_ACP_OAUTH2_MANAGE_PRIV_V1,
-            JSON_IDM_ACP_RADIUS_SECRET_READ_PRIV_V1,
-            JSON_IDM_ACP_RADIUS_SECRET_WRITE_PRIV_V1,
-            JSON_IDM_HP_ACP_SERVICE_ACCOUNT_INTO_PERSON_MIGRATE_V1,
-            JSON_IDM_ACP_OAUTH2_READ_PRIV_V1,
-            JSON_IDM_HP_ACP_SYNC_ACCOUNT_MANAGE_PRIV_V1,
-        ];
-
-        let res: Result<(), _> = idm_entries
-            .iter()
-            .try_for_each(|e_str| self.internal_migrate_or_create_str(e_str));
-        if res.is_ok() {
-            admin_debug!("initialise_idm -> result Ok!");
-        } else {
-            admin_error!(?res, "initialise_idm p3 -> result");
-        }
-        debug_assert!(res.is_ok());
-        res?;
-
-        let idm_entries = [
-            E_IDM_UI_ENABLE_EXPERIMENTAL_FEATURES.clone(),
-            E_IDM_ACCOUNT_MAIL_READ_PRIV.clone(),
-            E_IDM_ACP_ACCOUNT_MAIL_READ_PRIV_V1.clone(),
-        ];
-
-        let res: Result<(), _> = idm_entries
-            .into_iter()
-            .try_for_each(|entry| self.internal_migrate_or_create(entry));
-        if res.is_ok() {
-            admin_debug!("initialise_idm -> result Ok!");
-        } else {
-            admin_error!(?res, "initialise_idm p3 -> result");
-        }
-        debug_assert!(res.is_ok());
-        res?;
-
-        self.changed_schema.set(true);
-        self.changed_acp.set(true);
-
-        Ok(())
-    }
-
     #[instrument(level = "debug", name = "reload_schema", skip(self))]
     fn reload_schema(&mut self) -> Result<(), OperationError> {
         // supply entries to the writable schema to reload from.
@@ -1928,511 +1393,8 @@ impl<'a> QueryServerWriteTransaction<'a> {
 
 #[cfg(test)]
 mod tests {
-    use std::sync::Arc;
-    use std::time::Duration;
 
-    use kanidm_proto::v1::SchemaError;
-
-    use crate::credential::policy::CryptoPolicy;
-    use crate::credential::Credential;
-    use crate::event::{CreateEvent, DeleteEvent, ModifyEvent, ReviveRecycledEvent, SearchEvent};
     use crate::prelude::*;
-
-    #[qs_test]
-    async fn test_create_user(server: &QueryServer) {
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        let filt = filter!(f_eq("name", PartialValue::new_iname("testperson")));
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-
-        let se1 = unsafe { SearchEvent::new_impersonate_entry(admin.clone(), filt.clone()) };
-        let se2 = unsafe { SearchEvent::new_impersonate_entry(admin, filt) };
-
-        let mut e = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("class", Value::new_class("account")),
-            ("name", Value::new_iname("testperson")),
-            ("spn", Value::new_spn_str("testperson", "example.com")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson")),
-            ("displayname", Value::new_utf8s("testperson"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e.clone()]);
-
-        let r1 = server_txn.search(&se1).expect("search failure");
-        assert!(r1.is_empty());
-
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        let r2 = server_txn.search(&se2).expect("search failure");
-        debug!("--> {:?}", r2);
-        assert!(r2.len() == 1);
-
-        // We apply some member-of in the server now, so we add these before we seal.
-        e.add_ava("class", Value::new_class("memberof"));
-        e.add_ava("memberof", Value::Refer(UUID_IDM_ALL_PERSONS));
-        e.add_ava("directmemberof", Value::Refer(UUID_IDM_ALL_PERSONS));
-        e.add_ava("memberof", Value::Refer(UUID_IDM_ALL_ACCOUNTS));
-        e.add_ava("directmemberof", Value::Refer(UUID_IDM_ALL_ACCOUNTS));
-
-        let expected = unsafe { vec![Arc::new(e.into_sealed_committed())] };
-
-        assert_eq!(r2, expected);
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    #[qs_test]
-    async fn test_init_idempotent_schema_core(server: &QueryServer) {
-        {
-            // Setup and abort.
-            let mut server_txn = server.write(duration_from_epoch_now()).await;
-            assert!(server_txn.initialise_schema_core().is_ok());
-        }
-        {
-            let mut server_txn = server.write(duration_from_epoch_now()).await;
-            assert!(server_txn.initialise_schema_core().is_ok());
-            assert!(server_txn.initialise_schema_core().is_ok());
-            assert!(server_txn.commit().is_ok());
-        }
-        {
-            // Now do it again in a new txn, but abort
-            let mut server_txn = server.write(duration_from_epoch_now()).await;
-            assert!(server_txn.initialise_schema_core().is_ok());
-        }
-        {
-            // Now do it again in a new txn.
-            let mut server_txn = server.write(duration_from_epoch_now()).await;
-            assert!(server_txn.initialise_schema_core().is_ok());
-            assert!(server_txn.commit().is_ok());
-        }
-    }
-
-    #[qs_test]
-    async fn test_modify(server: &QueryServer) {
-        // Create an object
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let e2 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson2")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63932"))
-            ),
-            ("description", Value::new_utf8s("testperson2")),
-            ("displayname", Value::new_utf8s("testperson2"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e1.clone(), e2.clone()]);
-
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Empty Modlist (filter is valid)
-        let me_emp = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_pres("class")),
-                ModifyList::new_list(vec![]),
-            )
-        };
-        assert!(server_txn.modify(&me_emp) == Err(OperationError::EmptyRequest));
-
-        // Mod changes no objects
-        let me_nochg = unsafe {
-            ModifyEvent::new_impersonate_entry_ser(
-                JSON_ADMIN_V1,
-                filter!(f_eq("name", PartialValue::new_iname("flarbalgarble"))),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("description"),
-                    Value::from("anusaosu"),
-                )]),
-            )
-        };
-        assert!(server_txn.modify(&me_nochg) == Err(OperationError::NoMatchingEntries));
-
-        // Filter is invalid to schema - to check this due to changes in the way events are
-        // handled, we put this via the internal modify function to get the modlist
-        // checked for us. Normal server operation doesn't allow weird bypasses like
-        // this.
-        let r_inv_1 = server_txn.internal_modify(
-            &filter!(f_eq("tnanuanou", PartialValue::new_iname("Flarbalgarble"))),
-            &ModifyList::new_list(vec![Modify::Present(
-                AttrString::from("description"),
-                Value::from("anusaosu"),
-            )]),
-        );
-        assert!(
-            r_inv_1
-                == Err(OperationError::SchemaViolation(
-                    SchemaError::InvalidAttribute("tnanuanou".to_string())
-                ))
-        );
-
-        // Mod is invalid to schema
-        let me_inv_m = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_pres("class")),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("htnaonu"),
-                    Value::from("anusaosu"),
-                )]),
-            )
-        };
-        assert!(
-            server_txn.modify(&me_inv_m)
-                == Err(OperationError::SchemaViolation(
-                    SchemaError::InvalidAttribute("htnaonu".to_string())
-                ))
-        );
-
-        // Mod single object
-        let me_sin = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson2"))),
-                ModifyList::new_list(vec![
-                    Modify::Purged(AttrString::from("description")),
-                    Modify::Present(AttrString::from("description"), Value::from("anusaosu")),
-                ]),
-            )
-        };
-        assert!(server_txn.modify(&me_sin).is_ok());
-
-        // Mod multiple object
-        let me_mult = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_or!([
-                    f_eq("name", PartialValue::new_iname("testperson1")),
-                    f_eq("name", PartialValue::new_iname("testperson2")),
-                ])),
-                ModifyList::new_list(vec![
-                    Modify::Purged(AttrString::from("description")),
-                    Modify::Present(AttrString::from("description"), Value::from("anusaosu")),
-                ]),
-            )
-        };
-        assert!(server_txn.modify(&me_mult).is_ok());
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    #[qs_test]
-    async fn test_modify_assert(server: &QueryServer) {
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-
-        let t_uuid = Uuid::new_v4();
-        let r_uuid = Uuid::new_v4();
-
-        assert!(server_txn
-            .internal_create(vec![entry_init!(
-                ("class", Value::new_class("object")),
-                ("uuid", Value::Uuid(t_uuid))
-            ),])
-            .is_ok());
-
-        // This assertion will FAIL
-        assert!(matches!(
-            server_txn.internal_modify_uuid(
-                t_uuid,
-                &ModifyList::new_list(vec![
-                    m_assert("uuid", &PartialValue::Uuid(r_uuid)),
-                    m_pres("description", &Value::Utf8("test".into()))
-                ])
-            ),
-            Err(OperationError::ModifyAssertionFailed)
-        ));
-
-        // This assertion will PASS
-        assert!(server_txn
-            .internal_modify_uuid(
-                t_uuid,
-                &ModifyList::new_list(vec![
-                    m_assert("uuid", &PartialValue::Uuid(t_uuid)),
-                    m_pres("description", &Value::Utf8("test".into()))
-                ])
-            )
-            .is_ok());
-    }
-
-    #[qs_test]
-    async fn test_modify_invalid_class(server: &QueryServer) {
-        // Test modifying an entry and adding an extra class, that would cause the entry
-        // to no longer conform to schema.
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e1.clone()]);
-
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Add class but no values
-        let me_sin = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson1"))),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("class"),
-                    Value::new_class("system_info"),
-                )]),
-            )
-        };
-        assert!(server_txn.modify(&me_sin).is_err());
-
-        // Add multivalue where not valid
-        let me_sin = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson1"))),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("name"),
-                    Value::new_iname("testpersonx"),
-                )]),
-            )
-        };
-        assert!(server_txn.modify(&me_sin).is_err());
-
-        // add class and valid values?
-        let me_sin = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson1"))),
-                ModifyList::new_list(vec![
-                    Modify::Present(AttrString::from("class"), Value::new_class("system_info")),
-                    // Modify::Present("domain".to_string(), Value::new_iutf8("domain.name")),
-                    Modify::Present(AttrString::from("version"), Value::new_uint32(1)),
-                ]),
-            )
-        };
-        assert!(server_txn.modify(&me_sin).is_ok());
-
-        // Replace a value
-        let me_sin = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson1"))),
-                ModifyList::new_list(vec![
-                    Modify::Purged(AttrString::from("name")),
-                    Modify::Present(AttrString::from("name"), Value::new_iname("testpersonx")),
-                ]),
-            )
-        };
-        assert!(server_txn.modify(&me_sin).is_ok());
-    }
-
-    #[qs_test]
-    async fn test_delete(server: &QueryServer) {
-        // Create
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let e2 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson2")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63932"))
-            ),
-            ("description", Value::new_utf8s("testperson")),
-            ("displayname", Value::new_utf8s("testperson2"))
-        );
-
-        let e3 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson3")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63933"))
-            ),
-            ("description", Value::new_utf8s("testperson")),
-            ("displayname", Value::new_utf8s("testperson3"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e1.clone(), e2.clone(), e3.clone()]);
-
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Delete filter is syntax invalid
-        let de_inv =
-            unsafe { DeleteEvent::new_internal_invalid(filter!(f_pres("nhtoaunaoehtnu"))) };
-        assert!(server_txn.delete(&de_inv).is_err());
-
-        // Delete deletes nothing
-        let de_empty = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_eq(
-                "uuid",
-                PartialValue::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-000000000000"))
-            )))
-        };
-        assert!(server_txn.delete(&de_empty).is_err());
-
-        // Delete matches one
-        let de_sin = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_eq(
-                "name",
-                PartialValue::new_iname("testperson3")
-            )))
-        };
-        assert!(server_txn.delete(&de_sin).is_ok());
-
-        // Delete matches many
-        let de_mult = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_eq(
-                "description",
-                PartialValue::new_utf8s("testperson")
-            )))
-        };
-        assert!(server_txn.delete(&de_mult).is_ok());
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    #[qs_test]
-    async fn test_tombstone(server: &QueryServer) {
-        // First we setup some timestamps
-        let time_p1 = duration_from_epoch_now();
-        let time_p2 = time_p1 + Duration::from_secs(CHANGELOG_MAX_AGE * 2);
-        let time_p3 = time_p2 + Duration::from_secs(CHANGELOG_MAX_AGE * 2);
-
-        let mut server_txn = server.write(time_p1).await;
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-
-        let filt_i_ts = filter_all!(f_eq("class", PartialValue::new_class("tombstone")));
-
-        // Create fake external requests. Probably from admin later
-        // Should we do this with impersonate instead of using the external
-        let me_ts = unsafe {
-            ModifyEvent::new_impersonate_entry(
-                admin.clone(),
-                filt_i_ts.clone(),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("class"),
-                    Value::new_class("tombstone"),
-                )]),
-            )
-        };
-
-        let de_ts = unsafe { DeleteEvent::new_impersonate_entry(admin.clone(), filt_i_ts.clone()) };
-        let se_ts = unsafe { SearchEvent::new_ext_impersonate_entry(admin, filt_i_ts.clone()) };
-
-        // First, create an entry, then push it through the lifecycle.
-        let e_ts = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("9557f49c-97a5-4277-a9a5-097d17eb8317"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-
-        let ce = CreateEvent::new_internal(vec![e_ts]);
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        let de_sin = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_or!([f_eq(
-                "name",
-                PartialValue::new_iname("testperson1")
-            )])))
-        };
-        assert!(server_txn.delete(&de_sin).is_ok());
-
-        // Commit
-        assert!(server_txn.commit().is_ok());
-
-        // Now, establish enough time for the recycled items to be purged.
-        let mut server_txn = server.write(time_p2).await;
-        assert!(server_txn.purge_recycled().is_ok());
-
-        // Now test the tombstone properties.
-
-        // Can it be seen (external search)
-        let r1 = server_txn.search(&se_ts).expect("search failed");
-        assert!(r1.is_empty());
-
-        // Can it be deleted (external delete)
-        // Should be err-no candidates.
-        assert!(server_txn.delete(&de_ts).is_err());
-
-        // Can it be modified? (external modify)
-        // Should be err-no candidates
-        assert!(server_txn.modify(&me_ts).is_err());
-
-        // Can it be seen (internal search)
-        // Internal search should see it.
-        let r2 = server_txn
-            .internal_search(filt_i_ts.clone())
-            .expect("internal search failed");
-        assert!(r2.len() == 1);
-
-        // If we purge now, nothing happens, we aren't past the time window.
-        assert!(server_txn.purge_tombstones().is_ok());
-
-        let r3 = server_txn
-            .internal_search(filt_i_ts.clone())
-            .expect("internal search failed");
-        assert!(r3.len() == 1);
-
-        // Commit
-        assert!(server_txn.commit().is_ok());
-
-        // New txn, push the cid forward.
-        let server_txn = server.write(time_p3).await;
-
-        // Now purge
-        assert!(server_txn.purge_tombstones().is_ok());
-
-        // Assert it's gone
-        // Internal search should not see it.
-        let r4 = server_txn
-            .internal_search(filt_i_ts)
-            .expect("internal search failed");
-        assert!(r4.is_empty());
-
-        assert!(server_txn.commit().is_ok());
-    }
 
     #[qs_test]
     async fn test_name_to_uuid(server: &QueryServer) {
@@ -2757,343 +1719,4 @@ mod tests {
         // Commit.
     }
 
-    #[qs_test]
-    async fn test_modify_password_only(server: &QueryServer) {
-        let e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("class", Value::new_class("account")),
-            ("name", Value::new_iname("testperson1")),
-            (
-                "uuid",
-                Value::Uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            ),
-            ("description", Value::new_utf8s("testperson1")),
-            ("displayname", Value::new_utf8s("testperson1"))
-        );
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        // Add the entry. Today we have no syntax to take simple str to a credential
-        // but honestly, that's probably okay :)
-        let ce = CreateEvent::new_internal(vec![e1]);
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Build the credential.
-        let p = CryptoPolicy::minimum();
-        let cred = Credential::new_password_only(&p, "test_password").unwrap();
-        let v_cred = Value::new_credential("primary", cred);
-        assert!(v_cred.validate());
-
-        // now modify and provide a primary credential.
-        let me_inv_m = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("name", PartialValue::new_iname("testperson1"))),
-                ModifyList::new_list(vec![Modify::Present(
-                    AttrString::from("primary_credential"),
-                    v_cred,
-                )]),
-            )
-        };
-        // go!
-        assert!(server_txn.modify(&me_inv_m).is_ok());
-
-        // assert it exists and the password checks out
-        let test_ent = server_txn
-            .internal_search_uuid(uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
-            .expect("failed");
-        // get the primary ava
-        let cred_ref = test_ent
-            .get_ava_single_credential("primary_credential")
-            .expect("Failed");
-        // do a pw check.
-        assert!(cred_ref.verify_password("test_password").unwrap());
-    }
-
-    fn create_user(name: &str, uuid: &str) -> Entry<EntryInit, EntryNew> {
-        entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("person")),
-            ("name", Value::new_iname(name)),
-            ("uuid", Value::new_uuid_s(uuid).expect("uuid")),
-            ("description", Value::new_utf8s("testperson-entry")),
-            ("displayname", Value::new_utf8s(name))
-        )
-    }
-
-    fn create_group(name: &str, uuid: &str, members: &[&str]) -> Entry<EntryInit, EntryNew> {
-        let mut e1 = entry_init!(
-            ("class", Value::new_class("object")),
-            ("class", Value::new_class("group")),
-            ("name", Value::new_iname(name)),
-            ("uuid", Value::new_uuid_s(uuid).expect("uuid")),
-            ("description", Value::new_utf8s("testgroup-entry"))
-        );
-        members
-            .iter()
-            .for_each(|m| e1.add_ava("member", Value::new_refer_s(m).unwrap()));
-        e1
-    }
-
-    fn check_entry_has_mo(qs: &QueryServerWriteTransaction, name: &str, mo: &str) -> bool {
-        let e = qs
-            .internal_search(filter!(f_eq("name", PartialValue::new_iname(name))))
-            .unwrap()
-            .pop()
-            .unwrap();
-
-        e.attribute_equality("memberof", &PartialValue::new_refer_s(mo).unwrap())
-    }
-
-    #[qs_test]
-    async fn test_revive_advanced_directmemberships(server: &QueryServer) {
-        // Create items
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        let admin = server_txn.internal_search_uuid(UUID_ADMIN).expect("failed");
-
-        // Right need a user in a direct group.
-        let u1 = create_user("u1", "22b47373-d123-421f-859e-9ddd8ab14a2a");
-        let g1 = create_group(
-            "g1",
-            "cca2bbfc-5b43-43f3-be9e-f5b03b3defec",
-            &["22b47373-d123-421f-859e-9ddd8ab14a2a"],
-        );
-
-        // Need a user in A -> B -> User, such that A/B are re-adde as MO
-        let u2 = create_user("u2", "5c19a4a2-b9f0-4429-b130-5782de5fddda");
-        let g2a = create_group(
-            "g2a",
-            "e44cf9cd-9941-44cb-a02f-307b6e15ac54",
-            &["5c19a4a2-b9f0-4429-b130-5782de5fddda"],
-        );
-        let g2b = create_group(
-            "g2b",
-            "d3132e6e-18ce-4b87-bee1-1d25e4bfe96d",
-            &["e44cf9cd-9941-44cb-a02f-307b6e15ac54"],
-        );
-
-        // Need a user in a group that is recycled after, then revived at the same time.
-        let u3 = create_user("u3", "68467a41-6e8e-44d0-9214-a5164e75ca03");
-        let g3 = create_group(
-            "g3",
-            "36048117-e479-45ed-aeb5-611e8d83d5b1",
-            &["68467a41-6e8e-44d0-9214-a5164e75ca03"],
-        );
-
-        // A user in a group that is recycled, user is revived, THEN the group is. Group
-        // should be present in MO after the second revive.
-        let u4 = create_user("u4", "d696b10f-1729-4f1a-83d0-ca06525c2f59");
-        let g4 = create_group(
-            "g4",
-            "d5c59ac6-c533-4b00-989f-d0e183f07bab",
-            &["d696b10f-1729-4f1a-83d0-ca06525c2f59"],
-        );
-
-        let ce = CreateEvent::new_internal(vec![u1, g1, u2, g2a, g2b, u3, g3, u4, g4]);
-        let cr = server_txn.create(&ce);
-        assert!(cr.is_ok());
-
-        // Now recycle the needed entries.
-        let de = unsafe {
-            DeleteEvent::new_internal_invalid(filter!(f_or(vec![
-                f_eq("name", PartialValue::new_iname("u1")),
-                f_eq("name", PartialValue::new_iname("u2")),
-                f_eq("name", PartialValue::new_iname("u3")),
-                f_eq("name", PartialValue::new_iname("g3")),
-                f_eq("name", PartialValue::new_iname("u4")),
-                f_eq("name", PartialValue::new_iname("g4"))
-            ])))
-        };
-        assert!(server_txn.delete(&de).is_ok());
-
-        // Now revive and check each one, one at a time.
-        let rev1 = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin.clone(),
-                filter_all!(f_eq("name", PartialValue::new_iname("u1"))),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rev1).is_ok());
-        // check u1 contains MO ->
-        assert!(check_entry_has_mo(
-            &server_txn,
-            "u1",
-            "cca2bbfc-5b43-43f3-be9e-f5b03b3defec"
-        ));
-
-        // Revive u2 and check it has two mo.
-        let rev2 = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin.clone(),
-                filter_all!(f_eq("name", PartialValue::new_iname("u2"))),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rev2).is_ok());
-        assert!(check_entry_has_mo(
-            &server_txn,
-            "u2",
-            "e44cf9cd-9941-44cb-a02f-307b6e15ac54"
-        ));
-        assert!(check_entry_has_mo(
-            &server_txn,
-            "u2",
-            "d3132e6e-18ce-4b87-bee1-1d25e4bfe96d"
-        ));
-
-        // Revive u3 and g3 at the same time.
-        let rev3 = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin.clone(),
-                filter_all!(f_or(vec![
-                    f_eq("name", PartialValue::new_iname("u3")),
-                    f_eq("name", PartialValue::new_iname("g3"))
-                ])),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rev3).is_ok());
-        assert!(
-            check_entry_has_mo(&server_txn, "u3", "36048117-e479-45ed-aeb5-611e8d83d5b1") == false
-        );
-
-        // Revive u4, should NOT have the MO.
-        let rev4a = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin.clone(),
-                filter_all!(f_eq("name", PartialValue::new_iname("u4"))),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rev4a).is_ok());
-        assert!(
-            check_entry_has_mo(&server_txn, "u4", "d5c59ac6-c533-4b00-989f-d0e183f07bab") == false
-        );
-
-        // Now revive g4, should allow MO onto u4.
-        let rev4b = unsafe {
-            ReviveRecycledEvent::new_impersonate_entry(
-                admin,
-                filter_all!(f_eq("name", PartialValue::new_iname("g4"))),
-            )
-        };
-        assert!(server_txn.revive_recycled(&rev4b).is_ok());
-        assert!(
-            check_entry_has_mo(&server_txn, "u4", "d5c59ac6-c533-4b00-989f-d0e183f07bab") == false
-        );
-
-        assert!(server_txn.commit().is_ok());
-    }
-
-    #[qs_test_no_init]
-    async fn test_qs_upgrade_entry_attrs(server: &QueryServer) {
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        assert!(server_txn.upgrade_reindex(SYSTEM_INDEX_VERSION).is_ok());
-        assert!(server_txn.commit().is_ok());
-
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        server_txn.initialise_schema_core().unwrap();
-        server_txn.initialise_schema_idm().unwrap();
-        assert!(server_txn.commit().is_ok());
-
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        assert!(server_txn.upgrade_reindex(SYSTEM_INDEX_VERSION + 1).is_ok());
-        assert!(server_txn.commit().is_ok());
-
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        assert!(server_txn
-            .internal_migrate_or_create_str(JSON_SYSTEM_INFO_V1)
-            .is_ok());
-        assert!(server_txn
-            .internal_migrate_or_create_str(JSON_DOMAIN_INFO_V1)
-            .is_ok());
-        assert!(server_txn
-            .internal_migrate_or_create_str(JSON_SYSTEM_CONFIG_V1)
-            .is_ok());
-        assert!(server_txn.commit().is_ok());
-
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        // ++ Mod the schema to set name to the old string type
-        let me_syn = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_or!([
-                    f_eq("attributename", PartialValue::new_iutf8("name")),
-                    f_eq("attributename", PartialValue::new_iutf8("domain_name")),
-                ])),
-                ModifyList::new_purge_and_set(
-                    "syntax",
-                    Value::new_syntaxs("UTF8STRING_INSENSITIVE").unwrap(),
-                ),
-            )
-        };
-        assert!(server_txn.modify(&me_syn).is_ok());
-        assert!(server_txn.commit().is_ok());
-
-        let mut server_txn = server.write(duration_from_epoch_now()).await;
-        // ++ Mod domain name and name to be the old type.
-        let me_dn = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_eq("uuid", PartialValue::Uuid(UUID_DOMAIN_INFO))),
-                ModifyList::new_list(vec![
-                    Modify::Purged(AttrString::from("name")),
-                    Modify::Purged(AttrString::from("domain_name")),
-                    Modify::Present(AttrString::from("name"), Value::new_iutf8("domain_local")),
-                    Modify::Present(
-                        AttrString::from("domain_name"),
-                        Value::new_iutf8("example.com"),
-                    ),
-                ]),
-            )
-        };
-        assert!(server_txn.modify(&me_dn).is_ok());
-
-        // Now, both the types are invalid.
-
-        // WARNING! We can't commit here because this triggers domain_reload which will fail
-        // due to incorrect syntax of the domain name! Run the migration in the same txn!
-        // Trigger a schema reload.
-        assert!(server_txn.reload_schema().is_ok());
-
-        // We can't just re-run the migrate here because name takes it's definition from
-        // in memory, and we can't re-run the initial memory gen. So we just fix it to match
-        // what the migrate "would do".
-        let me_syn = unsafe {
-            ModifyEvent::new_internal_invalid(
-                filter!(f_or!([
-                    f_eq("attributename", PartialValue::new_iutf8("name")),
-                    f_eq("attributename", PartialValue::new_iutf8("domain_name")),
-                ])),
-                ModifyList::new_purge_and_set(
-                    "syntax",
-                    Value::new_syntaxs("UTF8STRING_INAME").unwrap(),
-                ),
-            )
-        };
-        assert!(server_txn.modify(&me_syn).is_ok());
-
-        // WARNING! We can't commit here because this triggers domain_reload which will fail
-        // due to incorrect syntax of the domain name! Run the migration in the same txn!
-        // Trigger a schema reload.
-        assert!(server_txn.reload_schema().is_ok());
-
-        // ++ Run the upgrade for X to Y
-        assert!(server_txn.migrate_2_to_3().is_ok());
-
-        assert!(server_txn.commit().is_ok());
-
-        // Assert that it migrated and worked as expected.
-        let server_txn = server.write(duration_from_epoch_now()).await;
-        let domain = server_txn
-            .internal_search_uuid(UUID_DOMAIN_INFO)
-            .expect("failed");
-        // ++ assert all names are iname
-        assert!(
-            domain.get_ava_set("name").expect("no name?").syntax() == SyntaxType::Utf8StringIname
-        );
-        // ++ assert all domain/domain_name are iname
-        assert!(
-            domain
-                .get_ava_set("domain_name")
-                .expect("no domain_name?")
-                .syntax()
-                == SyntaxType::Utf8StringIname
-        );
-        assert!(server_txn.commit().is_ok());
-    }
 }
